+++ server/src/service/routes/task/create.rs
@@ -66,7 +66,7 @@ pub async fn create(
         .err_with_type(accept)?;
 
     create_response_with(
-        TaskContainer::for_supplier(task),
+        TaskContainer::for_doctor(task),
         accept,
         StatusCode::CREATED,
         |_| (),